pyproject.toml
[project]
name = "language-arts-agent"
version = "0.1.0"
description = "English Language Arts Teaching App with AI-powered assessment"
authors = [
    {name = "Language Arts Team"}
]
dependencies = [
    "flask>=2.3.3",
    "flask-sqlalchemy>=3.0.5",
    "flask-migrate>=4.0.5",
    "flask-jwt-extended>=4.5.3",
    "flask-cors>=4.0.0",
    "flask-login>=0.6.3",
    "psycopg2-binary>=2.9.7",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "openai>=0.28.1",
    "azure-cognitiveservices-speech>=1.32.1",
    "pillow>=10.0.1",
    "boto3>=1.28.62",
    "gunicorn>=21.2.0",
    "pydub>=0.25.1",
]
readme = "README.md"
requires-python = ">= 3.8.1"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-flask>=1.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.2",
    "pytest-flask>=1.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
